int to long change missed in 699a1c

+++ src/main/java/de/catma/repository/git/interfaces/IRemoteGitManagerRestricted.java
@@ -49,7 +49,7 @@ public interface IRemoteGitManagerRestricted extends IGitUserInformation, ICommo
 	 * @throws IOException if something went wrong while deleting the remote
 	 *         repository
 	 */
-	void deleteRepository(int repositoryId) throws IOException;
+	void deleteRepository(long repositoryId) throws IOException;
 
 	/**
 	 * Creates a new remote group with the <code>name</code>, <code>path</code> and
